move fiat-convert to rpc - adjust imports
