+++ freqtrade/rpc/rpc.py
@@ -14,9 +14,9 @@ from numpy import mean, nan_to_num, NAN
 from pandas import DataFrame
 
 from freqtrade import TemporaryError, DependencyException
-from freqtrade.fiat_convert import CryptoToFiatConverter
 from freqtrade.misc import shorten_date
 from freqtrade.persistence import Trade
+from freqtrade.rpc.fiat_convert import CryptoToFiatConverter
 from freqtrade.state import State
 from freqtrade.strategy.interface import SellType
